Moved TransformProperty to common types
